.backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
requests==2.31.0
python-docx==1.1.0
PyPDF2==3.0.1
pydantic==2.5.0
# New agent system dependencies
asyncio
logging